refactor: rename cards to be more explicit
