+++ components/music/cards/AlbumCard/CompactAlbum.tsx
