+++ components/music/cards/AlbumCard/NormalAlbum.tsx
