+++ components/music/cards/index.tsx
@@ -1,4 +1,4 @@
-export { default as Album } from "./Album";
-export { default as Artist } from "./Artist";
-export { default as Playlist } from "./Playlist";
-export { default as Track } from "./Track";
+export { default as AlbumCard } from "./AlbumCard";
+export { default as ArtistCard } from "./Artist";
+export { default as PlaylistCard } from "./Playlist";
+export { default as TrackCard } from "./Track";
