+++ components/music/grids/TracksGrid.tsx
@@ -1,4 +1,4 @@
-import { Track } from "@/components/music/cards";
+import { TrackCard } from "@/components/music/cards";
 
 const TracksGrid = ({
   tracks,
@@ -17,7 +17,7 @@ const TracksGrid = ({
   >
     {tracks.map((track, index) => (
       // @ts-expect-error Next 13 handles async components
-      <Track key={track?.id ?? index} track={track} />
+      <TrackCard key={track?.id ?? index} track={track} />
     ))}
   </div>
 );
